COC-22-ApiCall: Added Retrofit & log the JSON WIP
